# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rubber Plantation Partnership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{clerkUserId}/role:
    put:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - in: path
          name: clerkUserId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserRoleInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{clerkUserId}/projects:
    post:
      operationId: assignUserToProject
      tags: [users]
      summary: Assign a user to a project
      parameters:
        - in: path
          name: clerkUserId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignProjectInput"
      responses:
        "200":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /users/{clerkUserId}/projects/{projectId}:
    delete:
      operationId: removeUserFromProject
      tags: [users]
      summary: Remove a user from a project assignment (admin only)
      parameters:
        - in: path
          name: clerkUserId
          required: true
          schema:
            type: string
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all plantation projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new plantation project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a plantation project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a plantation project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a plantation project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, location, district, state, landArea, landAreaUnit, status, startDate, termYears, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        village:
          type: ["string", "null"]
        district:
          type: string
        state:
          type: string
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: ["number", "null"]
        landValuePerUnit:
          type: ["number", "null"]
        status:
          type: string
          enum: [planning, developing, maturing, tapping, completed, suspended]
        startDate:
          type: string
        expectedMaturityDate:
          type: ["string", "null"]
        termYears:
          type: integer
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ProjectInput:
      type: object
      required: [name, location, district, state, landArea, landAreaUnit, status, startDate, termYears]
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        village:
          type: string
        district:
          type: string
        state:
          type: string
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: number
        landValuePerUnit:
          type: number
        status:
          type: string
          enum: [planning, developing, maturing, tapping, completed, suspended]
        startDate:
          type: string
        expectedMaturityDate:
          type: string
        termYears:
          type: integer
        notes:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        village:
          type: string
        district:
          type: string
        state:
          type: string
        landArea:
          type: number
        landAreaUnit:
          type: string
        landNotionalValue:
          type: number
        landValuePerUnit:
          type: number
        status:
          type: string
          enum: [planning, developing, maturing, tapping, completed, suspended]
        startDate:
          type: string
        expectedMaturityDate:
          type: string
        termYears:
          type: integer
        notes:
          type: string
...
    UserProfile:
      type: object
      required: [clerkUserId, role, assignedProjectIds]
      properties:
        clerkUserId:
          type: string
        role:
          type: string
          enum: [admin, developer, landowner, investor, employee, operational_staff]
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        assignedProjectIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
...
    SetUserRoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, developer, landowner, investor, employee, operational_staff]
    AssignProjectInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: string
          format: uuid
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean