# lib/api-zod/src/generated/types/assignProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rubber Plantation Partnership API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignProjectInput {
  projectId: string;
}